refactor(admin): split content analytics sections
Move the content analytics tables and integration cards out of the route module. The page now owns layout and suspense boundaries while pipeline velocity, series performance, Typefully, Instagram, and PostHog sections live in a focused server module.

Verified: pnpm exec prettier --check apps/admin/src/app/content/analytics/page.tsx apps/admin/src/app/content/analytics/analytics-sections.tsx

Verified: git diff --check -- apps/admin/src/app/content/analytics/page.tsx apps/admin/src/app/content/analytics/analytics-sections.tsx

Verified: pnpm --filter @anipotts/admin typecheck

Verified: pnpm --filter @anipotts/admin lint

Verified: pnpm --filter @anipotts/admin build

Verified: pnpm validate

Verified: pnpm dlx knip --no-exit-code
